setup/requirements.txt
Django==3.2
boto3==1.24.10
gunicorn==20.1.0
django-storages==1.12.3
django-simple-history==3.1.1
dj-database-url==0.5.0
Pillow==9.1.1
psycopg2==2.9.3

# REST
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0

# OpenAPI schema support
PyYAML==6.0
uritemplate==4.1.1

# utilities
requests==2.28.0
openpyxl==3.0.10
fpdf==1.7.2
qrcode==7.3.1

# captcha
django-simple-captcha==0.5.17

# wagtail==3.0.1
# psycopg2-binary==2.9.3
